refactor: move Space type to types folder
