+++ src/components/NewTab/Tabs/index.tsx
@@ -1,5 +1,5 @@
 import TabCard from "./TabCard";
-import { Space } from "../index";
+import { Space } from "../../../types/space";
 import { Tab } from "../../../types/tab";
 import { useLocation } from "react-router-dom";
