+++ src/types/space.ts
@@ -1,0 +1,9 @@
+export interface SpaceDoc {
+  title: string;
+  isArchived?: boolean;
+}
+
+export interface Space extends SpaceDoc {
+  id: string;
+  isEditing: boolean;
+}
